Let AVR timer_count increment in atomic block
This is a hacky way to do it, using a 16-bit timer in normal mode is
probably better since that would allow longer atomic blocks without any
timer function calls. But I'm not sure how to handle timer overrun since
the counter value needs to be divided by for example 250 for atmega32u4.
Not a multiple of 2 so unsigned overflow doesn't automatically take care
of things.

With this, at least use cases that continuously check for timeout won't
get stuck because interrupts are disabled.